Use size_t in place of unsigned

+++ include/skeleton_aligner/utils.h
@@ -56,7 +56,7 @@ tf2::Transform weightedAverage(const std::vector<tf2::Transform>& t_tfs,
                                const double& t_weight = 1.);
 tf2::Transform weightedAverage(const tf2::Transform& t_tf,
                                const tf2::Transform& t_prev_avg,
-                               const unsigned& t_n_elems,
+                               const size_t& t_n_elems,
                                const double& t_weight = 1.);
 
 }  // namespace utils
